+++ src/components/Select/types.ts
@@ -5,7 +5,13 @@ import {Option, OptionGroup} from './tech-components';
 
 export type SelectProps = Pick<ControlGroupProps, 'name' | 'disabled'> & {
     onUpdate?: (value: string[]) => void;
-    renderOption?: (option: SelectOption) => React.ReactNode;
+    onOpenChange?: (open: boolean) => void;
+    renderControl?: (props: {
+        onClick: (e: React.MouseEvent<HTMLElement>) => void;
+        onKeyDown: (e: React.KeyboardEvent<HTMLElement>) => void;
+        ref: React.Ref<HTMLElement>;
+    }) => React.ReactElement;
+    renderOption?: (option: SelectOption) => React.ReactElement;
     getOptionHeight?: (option: SelectOption) => number;
     view?: TextInputView;
     size?: TextInputSize;